+++ packages/base/CHANGELOG.md
@@ -14,6 +14,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Removed
 
 ### Fixed
+ - **BREAKING** `SyncConnection.insert` return numeric value (instead of array) ONLY in case fieldList is === `['ID']` (as documented). 
+   Before this fix numeric value is returned for any fieldList with ONE element.  
 
 ## [5.20.8] - 2021-07-08
 ## [5.20.7] - 2021-05-24